+++ SFMLEngine/src/Tilemap.cpp
@@ -1,5 +1,8 @@
 #include "Tilemap.h"
 #include "Components/SpriteRenderer.h"
+#include "LifeCycleManager.h"
+#include "Entity.h"
+
 int32_t SFENG::Tilemap::s_TileNum = 0;
 
 SFENG::Tilemap::Tilemap(const Vec2i &size, const Vec2f &startPos, const Vec2f &tileSize, const std::string &spriteSheet, LifeCycleManager *LCM)
@@ -23,22 +26,22 @@ void SFENG::Tilemap::ReadMap(const std::string &map)
 	if (file.is_open())
 	{
 		sf::IntRect rect;
-		rect.width = m_TileSize.x;
-		rect.height = m_TileSize.y;
+		rect.width = (int)m_TileSize.x;
+		rect.height = (int)m_TileSize.y;
 		for (int i = 0; i < m_Size.y; i++)	   // Rows
 			for (int j = 0; j < m_Size.x; j++) // Columns
 			{
 				int32_t index = j + i * m_Size.x;
 				std::string str;
 				file >> str;
-				int in = str.find("_");
+				size_t in = str.find("_");
 				if (in != std::string::npos)
 				{
 					uint8_t x = std::stoi(str.substr(0, in));
 					uint8_t y = std::stoi(str.substr(in + 1));
 					m_Tiles[index]->trans->position = (Vec2f(m_StartPos.y + j * m_TileSize.x, m_StartPos.y + i * m_TileSize.y));
-					rect.left = x * m_TileSize.x;
-					rect.top = y * m_TileSize.y;
+					rect.left =(int) (x * m_TileSize.x);
+					rect.top =(int) (y * m_TileSize.y);
 					m_Tiles[index]->sr->SetRect(rect);
 				}
 			}
